+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 2024-09-26
+
+### Added 
+
+- Definition method for king, queen, bishop and knight
+- Tests fro piece str one for each piece
+
+### Changed
+
+- Pawns possible_positions_move method now only have black and white options
+
 ## [0.1.5] - 2024-09-24
 
 ### Added